backend/app/requirements.txt
fastapi>=0.95
uvicorn[standard]>=0.20
asyncpg>=0.27
pydantic>=1.10
python-dotenv>=1.0
bcrypt>=4.0
PyJWT>=2.8
PyGithub>=1.59
